Move classes into their own file from servlets
