+++ portfolio/src/main/java/com/google/sps/servlets/LogInServlet.java
@@ -29,32 +29,6 @@ import javax.servlet.annotation.WebServlet;
 @WebServlet("/loggedIn")
 public class LogInServlet extends HttpServlet {
 
-        public class LoggedInStatus {
-        private boolean loggedIn;
-        private String email;
-        private String logoutUrl;
-        private String loginUrl;
-
-        public LoggedInStatus(boolean loggedIn) {
-            this.loggedIn = loggedIn;
-            this.email = email;
-            this.loginUrl = loginUrl;
-            this.logoutUrl = logoutUrl;
-        }
-
-        public void setEmail(String email){
-            this.email = email;
-        }
-
-        public void setLogoutUrl(String logoutUrl){
-            this.logoutUrl = logoutUrl;
-        }
-
-        public void setLoginUrl(String loginUrl){
-            this.loginUrl = loginUrl;
-        }
-    }
-
     @Override
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
